gen-val/src/generation/src/NIST.CVP.ACVTS.Libraries.Generation/SHA2/v1_0/TestGroupGeneratorFactory.cs
using System.Collections.Generic;
using NIST.CVP.ACVTS.Libraries.Generation.Core;

namespace NIST.CVP.ACVTS.Libraries.Generation.SHA2.v1_0
{
    public class TestGroupGeneratorFactory : ITestGroupGeneratorFactory<Parameters, TestGroup, TestCase>
    {
        public IEnumerable<ITestGroupGeneratorAsync<Parameters, TestGroup, TestCase>> GetTestGroupGenerators(Parameters parameters)
        {
            var list = new HashSet<ITestGroupGeneratorAsync<Parameters, TestGroup, TestCase>>
                {
                    new TestGroupGeneratorAlgorithmFunctionalTest(),
                    new TestGroupGeneratorMonteCarloTest(),
                    new TestGroupGeneratorLargeDataTest()
                };

            return list;
        }
    }
}
